Fixed CObjPointSortArray to accept a template argument inheriting from CPointSort instead of a un-templated CPointSort-deriving class. This is important because CSectorTemplate::m_Teleports is a CObjPointSortArray which holds CTeleport objects, which inherits from CPointSort. We want a CObjPointSortArray that can also directly store a CTeleport class. CPointSort inherits from CPointMap and CPointBase but, for reasons highlighted in a comment in the source, by design those class are not virtual. This means that there's not a virtual destructor and that when CObjPointSortArray goes out of scope/is destroyed, it calls the destructor of CPointSort, not the one of the topmost class (CTeleport). This leads to a memory leak.
Renamed CResourceBase to CResourceHolder, which is far more in line with its function.
Moved CStartLoc and CTeleport classes to their own source files.

+++ src/common/CRect.cpp
@@ -214,7 +214,7 @@ tchar * CRect::Write( tchar * pBuffer, uint uiBufferLen) const
 lpctstr CRect::Write() const
 {
 	ADDTOCALLSTACK("CRect::Write");
-	return Write( Str_GetTemp(), Str_TempLength() );
+	return Write( Str_GetTemp(), (uint)Str_TempLength() );
 }
 
 CPointBase CRect::GetCenter() const
